Fix implicit conversion warning in FillCrossHatch.cpp

+++ src/libslic3r/Fill/FillCrossHatch.cpp
@@ -220,7 +220,7 @@ void FillCrossHatch ::_fill_surface_single(
     }
 
     if (!polylines.empty()) {
-        int infill_start_idx = polylines_out.size(); // only rotate what belongs to us.
+        int infill_start_idx = static_cast<int>(polylines_out.size()); // only rotate what belongs to us.
         // connect lines
         chain_or_connect_infill(std::move(polylines), expolygon, polylines_out, this->spacing, params);
